Extract capture_exception method that's specific for web requests
