+++ sentry-ruby/lib/sentry/rack/capture_exception.rb
@@ -27,12 +27,12 @@ module Sentry
           rescue Sentry::Error
             raise # Don't capture Sentry errors
           rescue Exception => e
-            capture_exception(e, env)
+            Sentry::Rack::capture_exception(e, env)
             raise
           end
 
           error = collect_exception(env)
-          capture_exception(error, env) if error
+          Sentry::Rack.capture_exception(error, env) if error
 
           response
         end
@@ -40,18 +40,6 @@ module Sentry
 
       def collect_exception(env)
         env['rack.exception'] || env['sinatra.error']
-      end
-
-      def capture_exception(exception, env, **options)
-        if requested_at = env['sentry.requested_at']
-          options[:time_spent] = Time.now - requested_at
-        end
-
-        Sentry.capture_exception(exception, **options) do |evt|
-          evt.interface :http do |int|
-            int.from_rack(env)
-          end
-        end
       end
     end
   end
